requirements.txt
requests
pydantic
boto3
sqlalchemy
psycopg2-binary
alembic
python-dotenv
fastapi
httpx
uvicorn